Polish BENCHMARK.md and trim JMH suite for 1.7.0
- Re-measure Read / Write / Sustained / SharedStrings on 1.7.0-SNAPSHOT
- Sort all tables layer-grouped: jackson / FastExcel / Apache Fesod / Apache POI / jackson-spreadsheet (POI User Model)
- Drop 500K column from flat tables; remove @Param "500000" from ReadBenchmark / WriteBenchmark / SharedStringsBenchmark
- Rename BackWriteBenchmark -> LargeInnerListBenchmark (add read symmetry, anchor=true on Order.id)
- Rename NestedIntParityBenchmark -> NestedListBenchmark; mark both internal-only in javadoc
- Remove NestedRead/Write/SustainedThroughput, CellTypeOverhead, StringOnly, mixed-type NestedCellParity, NestedStringParity benchmarks and their fixtures
