+++ src/jmh/java/io/github/scndry/jackson/dataformat/spreadsheet/WriteBenchmark.java
@@ -25,7 +25,7 @@ import org.openjdk.jmh.infra.Blackhole;
 @Fork(1)
 public class WriteBenchmark {
 
-    @Param({"1000", "10000", "50000", "100000", "500000"})
+    @Param({"1000", "10000", "50000", "100000"})
     int rowCount;
 
     List<BenchRow> data;
